correct reader device index, test=develop
